mcp-agent/requirements.txt
# Core LLM Framework
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0

# MCP Protocol
mcp>=1.0.0

# Configuration
python-dotenv>=1.0.0